Added Roboto font as dependency instead of link, files refactor
