+++ package.json
@@ -17,11 +17,10 @@
     "@material-ui/core": "^4.9.12",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/styles": "^4.9.10",
-    "gatsby": "^2.19.5",
+    "gatsby": "^2.21.10",
     "gatsby-plugin-canonical-urls": "^2.3.0",
     "gatsby-plugin-google-analytics": "^2.3.0",
     "gatsby-plugin-manifest": "^2.4.1",
-    "gatsby-plugin-material-ui": "^2.1.8",
     "gatsby-plugin-offline": "^3.0.41",
     "gatsby-plugin-react-helmet": "^3.3.0",
     "gatsby-plugin-robots-txt": "^1.5.0",
@@ -32,7 +31,8 @@
     "normalize.css": "^8.0.1",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
-    "react-helmet": "^6.0.0"
+    "react-helmet": "^6.0.0",
+    "typeface-roboto": "0.0.75"
   },
   "devDependencies": {
     "dotenv": "^8.2.0",
